Add replica of Vive hinge interface
This was the original part I was planning to use to interface the
headset to the headgear, a clone of the mechanism used by the stock
parts to attach to the headset.

I had mothballed this idea after printing it in both PLA and ABS and
finding that neither material was strong enough to support this idea at
the sizes involved; however, I've just tried this idea again with PETG
and it seems promising, so I'm going to try re-integrating it.